requirements.txt
langchain
langchain-openai
google-search-results
wikipedia
beautifulsoup4
requests
python-dotenv
